==> 04_extract_selected_cities.ipynb <==
Firstly, we decided we wanted to narrow our focus down to the United States. In order to do that, we created a new data set where we copied only the lines that included information on the United States.

import pandas as pd

# Specify the file path to the CSV file
file_path = "city_temperature.csv"

# Read the CSV file into a DataFrame
df = pd.read_csv(file_path, dtype={
    '1': str,
    '2': str,
    '3': str,
    '4': str,
}, low_memory=False)
df.columns


The following code was used to create the final dataset we wanted to use. It takes the huge dataset of daily temperatures and narrows the data down to the cities we wanted. It also included a line that checks for empty data values that are marked by 0's or -99's.

import pandas as pd

# Specify the file path to the CSV file
file_path = "C:\\Users\\Bryan\\Downloads\\city_temperature.csv"

# Read the CSV file into a DataFrame
df = pd.read_csv(file_path, dtype={'1': str, '2': str, '3': str, '4': str, '5': int, '6': int, '7': int, '8': float}, 
                 low_memory=False)

# Filter data for specific cities
filtered_df = df[(df['City'].isin(['Grand Junction', 'Pueblo', 'Salt Lake City', 'San Angelo', 'Denver', 'Zagreb', 'Milan', 
                                   'Belgrade', 'Madrid', 'Bishkek']))]
#Data cleaning for null values
filtered_df_1 = filtered_df[((filtered_df['Day'] != 0) & (filtered_df['AvgTemperature'] != -99))]

# Specify the file path for the new file
new_file_path = "glob_city_temperatures.csv"

# Save the filtered data to a new CSV file
filtered_df_1.to_csv(new_file_path, index=False)
