==> 05_selected_cities_avg_std.ipynb <==
The following code intends to create a CSV file with data that includes both the monthly average temperatures and the standard deviations of each month.

import numpy as np
import pandas as pd

# Read the CSV file into a DataFrame
df = pd.read_csv(r'glob_city_temperatures.csv') 

# Define data types for columns
data_types = {'City': str, 'Month': int, 'Day': int, 'Year': int, 'AvgTemperature': float}

# Drop unnecessary columns
df = df.drop(['State', 'Region', 'Country'], axis=1)

# Filter the DataFrame for the year 2019
filtered_df = df[(df['Year'] == 2019)]

# Calculate the monthly averages of temperatures for each city
monthly_averages = filtered_df.groupby(['Month','City'])['AvgTemperature'].mean().reset_index()

# Calculate the monthly standard deviations of temperatures for each city
monthly_std = filtered_df.groupby(['Month','City'])['AvgTemperature'].std().reset_index()

# Merge the average and standard deviation DataFrames
combined_df = monthly_averages.merge(monthly_std, on=['Month', 'City'], suffixes=('_avg', '_std'))

# Save the combined DataFrame to a CSV file
combined_df.to_csv('monthly_averages&std.csv', index=False)

The following code intends to return the standard deviations in August in ascending order.

import numpy as np
import pandas as pd

# Read the CSV file into a DataFrame
df = pd.read_csv(r'glob_city_temperatures.csv')

# Define data types for columns
data_types = {'City': str, 'Month': int, 'Day': int, 'Year': int, 'AvgTemperature': float}

# Drop unnecessary columns
df = df.drop(['State', 'Region', 'Country'], axis=1)

# Filter the DataFrame for the year 2019 and August
filtered_df = df[(df['Year'] == 2019) & (df['Month'].isin([8]))]

# Calculate the monthly averages of temperatures for each city
monthly_averages = filtered_df.groupby(['Month', 'City'])['AvgTemperature'].mean()

# Calculate the monthly standard deviations of temperatures for each city
monthly_std = filtered_df.groupby(['Month', 'City'])['AvgTemperature'].std()

# Sort the standard deviations in ascending order
sorted_std_df = monthly_std.sort_values(ascending=True)

# Print the sorted DataFrame
sorted_std_df
